TaskGenerateService and TaskExecuteService

+++ exchangis-job/exchangis-job-launcher/src/main/java/com/webank/wedatasphere/exchangis/job/launcher/domain/task/TaskProgressInfo.java
@@ -82,7 +82,7 @@ public class TaskProgressInfo {
         this.status = status;
     }
 
-    public double getProgress() {
+    public float getProgress() {
         return progress;
     }
